pyproject.toml
[build-system]
requires = [ "poetry>=1.0.0",]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "maintenance_app"
version = "0.1.0"
description = "Library providing metrics for maintenance"
authors = ["Ksenia Bolshakova", ]
license = "Proprietary"
include = [ "py.typed",]
packages = [{include = "maintenance_app"}]

[[tool.poetry.source]]
name = "private-pypi"
url = "https://pypi.int.quantlane.com/pypi/"

[[tool.poetry.source]]
name = "pypi-cache"
url = "https://pypi-cache.int.quantlane.com/root/pypi/+simple/"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.5"
radon = "^6.0.1"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
coverage = "^7.2.6"
python-gitlab = "^3.15.0"
asyncpg = "^0.28.0"
python-decouple = "^3.6"
logwood = "^3.6.0"
